synchronize with skeleton

+++ Vagrantfile
@@ -1,23 +1,28 @@
+def determine_bridge
+  if RbConfig::CONFIG['host_os'] =~ /mswin32|mingw32/
+    'Ethernet'
+  else
+    'eth0'
+  end
+end
+
+def read_or_determine_bridge
+  if File.exist?('tmp/ethernet-device.txt')
+    File.read('tmp/ethernet-device.txt').chomp
+  else
+    bridge = determine_bridge
+    File.write('tmp/ethernet-device.txt', bridge + "\n")
+    bridge
+  end
+end
+
 Vagrant.configure('2') do |config|
   config.vm.box = 'debian/stretch64'
   Dir.mkdir('tmp') unless File.exist?('tmp')
-
-  if File.exist?('tmp/ethernet-device.txt')
-    bridge = File.read('tmp/ethernet-device.txt').chomp
-  else
-    if RbConfig::CONFIG['host_os'] =~ /mswin32|mingw32/
-      bridge = 'Ethernet'
-    else
-      bridge = 'eth0'
-    end
-
-    File.write('tmp/ethernet-device.txt', bridge + "\n")
-  end
-
-  if not File.exist?('tmp/hostname.txt')
-    File.write('tmp/hostname.txt', "rs\n")
-  end
-
+  File.write('tmp/hostname.txt', "rs\n") unless File.exist?(
+    'tmp/hostname.txt'
+  )
+  bridge = read_or_determine_bridge
   config.vm.network :public_network, bridge: bridge
   config.vm.network :private_network, ip: '192.168.42.3'
